legacy/pyproject.toml
[project]
name = "service-fusion-tools"
version = "0.1.0"
description = "Tools for extracting data objects from the Service Fusion API spec."
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.0.0",
  "pyyaml>=6.0.2",
  "ruamel.yaml>=0.18.6",
  "typeguard>=4.0.0"
]

[project.optional-dependencies]
runtime = [
]

[dependency-groups]
dev = []
